Add Tab Navigation styles and component for Agent and Chat modes
